requirements.txt
# Bağımlılıklar - Fırat Mevzuat RAG Sistemi

# PDF İşleme
pdfplumber==0.11.0
PyMuPDF==1.24.5

# NLP & Embedding (BERTurk)
transformers==4.41.2
torch==2.3.0
sentence-transformers==3.0.1

# Vektör Veritabanı
chromadb==0.5.3

# LLM API
openai==1.35.3
google-generativeai==0.7.2

# API Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Değerlendirme & Metrikler
rouge-score==0.1.2
nltk==3.8.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4

# Yardımcı
tqdm==4.66.4
httpx==0.27.0
pydantic==2.7.4
